Corrected def of czero, cone, ci as real
This is a pretty nasty bug that actually happens every once in a while
in real life as well. Unfortunately, the compiler doesn't complain if
you define a real number as asign it a complex number. This is
especially bad for ci = cmplx(zero, one, idp) which will be simply zero
if ci is defined as real.